+++ lib/karafka/pro/connection/manager.rb
@@ -17,46 +17,36 @@ module Karafka
       # Manager that can handle working with multiplexed connections.
       #
       # This manager takes into consideration the number of partitions assigned to the topics and
-      # does it best to balance. Additional connections may not always be utilized because
+      # does its best to balance. Additional connections may not always be utilized because
       # alongside of them, other processes may "hijack" the assignment. In such cases those extra
       # empty connections will be turned off after a while.
       #
       # @note Manager operations relate to consumer groups and not subscription groups. Since
       #   cluster operations can cause consumer group wide effects, we always apply only one
-      #   change on a consumer group
+      #   change on a consumer group.
+      #
+      # @note Since we collect statistical data from listeners and this happens in a background
+      #   thread, we need to make sure we lock not to have race conditions with expired data
+      #   eviction.
       class Manager < Karafka::Connection::Manager
         include Core::Helpers::Time
 
         # How long should we wait after a rebalance before doing anything on a consumer group
         #
-        # @param initial_delay [Integer] how long should we wait after start of the process, before
-        #   we take any scaling action. During deployments it may take the whole cluster a moment
-        #   to reach a stable state, hence we delay that independetly from making intermediate
-        #   decisions
-        # @param scale_down_delay [Integer] lag between last known change on a given consumer group
-        #   and the moment we can scale down. Since we scale down only empty connections, this
-        #   can be shorter than the scale up delay.
-        # @param scale_up_delay [Integer] Delay in-between last change on a consumer group and
-        #   the moment we can scale up the listeners. Since scaling up may cause revocation of
-        #   some partitions, we want to do it slowly not to cause issues in the cluster. Depending
-        #   on the type of work and its time it may be desirable to have it longer or shorter.
-        def initialize(
-          initial_delay: 5 * 60 * 10_0,
-          scale_down_delay: 60 * 10_0,
-          scale_up_delay: 5 * 60 * 10_0
-        )
+        # @param scale_delay [Integer] How long should we wait before making any changes. Any
+        #   change related to this consumer group will postpone the scaling operations. This is
+        #   done that way to prevent too many friction in the cluster.
+        def initialize(scale_delay: 5 * 60 * 10.0)
           super()
-          @initial_delay = initial_delay
-          @scale_down_delay = scale_down_delay
-          @scale_up_delay = scale_up_delay
-          @started_at = monotonic_now
+          @scale_delay = scale_delay
           @mutex = Mutex.new
           @changes = Hash.new do |h, k|
             h[k] = {
               state: '',
               join_state: '',
               state_age: 0,
-              state_age_sync: monotonic_now
+              state_age_sync: monotonic_now,
+              change_age: 0
             }
           end
         end
@@ -70,85 +60,63 @@ module Karafka
         def register(listeners)
           @listeners = listeners
 
-          in_sg_families do |subscription_group, sg_listeners|
-            consumer_group = sg_listeners.first.subscription_group.consumer_group
-            subscription_group = sg_listeners.first.subscription_group
-
-            multiplexing = subscription_group.multiplexing
-
+          in_sg_families do |multiplexing, sg_listeners|
             if multiplexing.active? && multiplexing.dynamic?
-              # Start absolute minimum that is expected in this multiplexed group for now and leave
-              # the upscaling for later. This will minimize number of connections in total in
-              # all consumer groups effectively allowing for faster start
-              #
-              # Once we reach a stable state, we can then upscale/downcase slowly as needed
-              sg_listeners.each(&:start!)
+              # Start as many boot listeners as user wants. If not configured, starts half of max.
+              sg_listeners.first(multiplexing.boot).each(&:start!)
             else
               sg_listeners.each(&:start!)
             end
           end
         end
 
+        # Collects data from the statistics about given subscription group. This is used to ensure
+        # that we do not rescale short after rebalances, deployments, etc.
+        # @param subscription_group_id [String] id of the subscription group for which statistics
+        #   were emitted
+        # @param statistics [Hash] emitted statistics
+        #
+        # @note Please note that while we collect here per subscription group, we use those metrics
+        #   collectively on a whole consumer group. This reduces the friction.
         def notice(subscription_group_id, statistics)
           @mutex.synchronize do
             times = []
-
-            times << statistics['brokers'].values.map { |stats| stats['stateage'] }.min / 1000
+            # stateage is in microseconds
+            # We monitor broker changes to make sure we do not introduce extra friction
+            times << statistics['brokers'].values.map { |stats| stats['stateage'] }.min / 1_000
             times << statistics['cgrp']['rebalance_age']
             times << statistics['cgrp']['stateage']
 
-            @changes[subscription_group_id][:state_age] = times.min
-            @changes[subscription_group_id][:join_state] = statistics['cgrp']['join_state']
-            @changes[subscription_group_id][:state] = statistics['cgrp']['state']
-            @changes[subscription_group_id][:state_age_sync] = monotonic_now
+            # Keep the previous change age for changes that were triggered by us
+            previous_change_age = @changes[subscription_group_id][:change_age]
+
+            @changes[subscription_group_id] = {
+              state_age: times.min,
+              change_age: previous_change_age,
+              join_state: statistics['cgrp']['join_state'],
+              state: statistics['cgrp']['state'],
+              state_age_sync: monotonic_now
+            }
           end
         end
 
+        # Indicates, that something has changed on a subscription group. We consider every single
+        # change we make as a change to the setup as well.
+        # @param subscription_group_id [String]
         def touch(subscription_group_id)
           @mutex.synchronize do
-            @changes[subscription_group_id][:state_age] = 0
+            @changes[subscription_group_id][:change_age] = 0
             @changes[subscription_group_id][:state_age_sync] = monotonic_now
-
-            @changes.delete_if do |k, v|
-              monotonic_now - v[:state_age_sync] >= 30_000
-            end
           end
         end
 
         # Shuts down all the listeners when it is time (including moving to quiet) or rescales
         # when it is needed
         def control
-          @mutex.synchronize do
-            @changes.delete_if do |k, v|
-              monotonic_now - v[:state_age_sync] >= 30_000
-            end
-          end
-
-          p 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx'
-
           Karafka::App.done? ? shutdown : rescale
         end
 
         private
-
-        # Handles two scenarios:
-        #   - Selects subscriptions that could benefit from having more parallel connections
-        #     to kafka and then upscales them
-        #   - Selects subscriptions that are idle (have nothing subscribed to them) and then shuts
-        #     them down
-        def rescale
-          #return
-          p 'aaaa'
-          # Do not take any actions if process is not alive for enough time
-          p [monotonic_now - @started_at, @initial_delay]
-          #return if monotonic_now - @started_at < @initial_delay
-
-          p 'uuuuuuuuuuuuu'
-
-          scale_down
-
-          scale_up
-        end
 
         # Handles the shutdown and quiet flows
         def shutdown
@@ -172,16 +140,10 @@ module Karafka
             Karafka::App.quieted!
           end
 
-          p Rdkafka::Config.opaques
-
-          p 'h1'
-
           return if Karafka::App.quiet?
 
-          p 'h2'
           @stopped_subscription_groups ||= Set.new
 
-          p 'h3'
           # Since separate subscription groups are subscribed to different topics, there is no risk
           # in shutting them down independently even if they operate in the same subscription group
           in_sg_families do |subscription_group, sg_listeners|
@@ -198,9 +160,6 @@ module Karafka
             active_sg_listeners.each(&:stop!)
           end
 
-          p 'h4'
-          p @listeners.map(&:status)
-
           return unless @listeners.active.all?(&:stopped?)
 
           # All listeners including pending need to be moved at the end to stopped state for
@@ -208,17 +167,30 @@ module Karafka
           @listeners.each(&:stop!)
         end
 
+        # Handles two scenarios:
+        #   - Selects subscriptions that could benefit from having more parallel connections
+        #     to kafka and then upscales them
+        #   - Selects subscriptions that are idle (have nothing subscribed to them) and then shuts
+        #     them down
+        #
+        # We always run scaling down and up because it may be applicable to different CGs
+        def rescale
+          evict
+
+          scale_down
+          scale_up
+        end
+
+        # Checks for connections without any assignments and scales them down.
+        # Does that only for dynamically multiplexed subscription groups
         def scale_down
-          p 'scaling down check...'
           sgs_in_use = Karafka::App.assignments.keys.map(&:subscription_group).uniq
 
           # Select connections for scaling down
           in_sg_families do |subscription_group_name, sg_listeners|
-            p stable?(sg_listeners)
             next unless stable?(sg_listeners)
 
             subscription_group = sg_listeners.first.subscription_group
-            consumer_group =  sg_listeners.first.subscription_group.consumer_group
             multiplexing = subscription_group.multiplexing
 
             next unless multiplexing.active?
@@ -232,6 +204,8 @@ module Karafka
               # Skip listeners that are already in standby
               next unless sg_listener.active?
 
+              touch(sg_listener.subscription_group.id)
+
               # Shut down not used connection
               sg_listener.stop!
 
@@ -240,37 +214,22 @@ module Karafka
           end
         end
 
-        def stable?(sg_listeners)
-          p 'test'
-
-          sg_listeners.all? do |sg_listener|
-            next true unless @changes.key?(sg_listener.subscription_group.id)
-
-            state = @changes.fetch(sg_listener.subscription_group.id)
-
-            p state
-
-            state[:state_age] >= @scale_up_delay &&
-              state[:state] == 'up' &&
-              state[:join_state] == 'steady'
-          end
-        end
-
+        # Checks if we have space to scale and if there are any assignments with multiple topics
+        # partitions assigned in sgs that can be scaled. If that is the case, we scale up.
         def scale_up
           multi_part_sgs_families = Karafka::App
-                             .assignments
-                             .select { |_, partitions| partitions.size > 1 }
-                             .keys
-                             .map(&:subscription_group)
-                             .map(&:name)
-                             .uniq
+                                    .assignments
+                                    .select { |_, partitions| partitions.size > 1 }
+                                    .keys
+                                    .map(&:subscription_group)
+                                    .map(&:name)
+                                    .uniq
 
           # Select connections for scaling up
           in_sg_families do |subscription_group_name, sg_listeners|
             next unless stable?(sg_listeners)
 
             subscription_group = sg_listeners.first.subscription_group
-            consumer_group = sg_listeners.first.subscription_group.consumer_group
             multiplexing = subscription_group.multiplexing
 
             next unless multiplexing.active?
@@ -283,8 +242,6 @@ module Karafka
               # Skip already active connections
               next unless sg_listener.pending? || sg_listener.stopped?
 
-
-              p 'startinggggggggggggggggggggggggggggggg'
               touch(sg_listener.subscription_group.id)
               sg_listener.start!
 
@@ -293,12 +250,44 @@ module Karafka
           end
         end
 
-        def too_early?(consumer_group, delay)
-          (monotonic_now - @changes[consumer_group.id]) < delay
+        # @param sg_listeners [Array<Listener>] listeners from one multiplexed sg
+        # @return [Boolean] is given subscription group listeners set stable. It is considered
+        #   stable when it had no changes happening on it recently and all relevant states in it
+        #   are also stable. This is a strong indicator that no rebalances or other operations are
+        #   happening at a given moment.
+        def stable?(sg_listeners)
+          # If none of listeners has changes reported it means we did not yet start collecting
+          # metrics about any of them and at least one must be present. We do not consider it
+          # stable in such case as we still are waiting for metrics.
+          return false if sg_listeners.none? do |sg_listener|
+            @changes.key?(sg_listener.subscription_group.id)
+          end
+
+          sg_listeners.all? do |sg_listener|
+            # Not all SGs may be started initially or may be stopped, we ignore them here as they
+            # are irrelevant from the point of view of establishing stability
+            next true unless @changes.key?(sg_listener.subscription_group.id)
+
+            state = @changes[sg_listener.subscription_group.id]
+
+            state[:state_age] >= @scale_delay &&
+              state[:change_age] >= @scale_delay &&
+              state[:state] == 'up' &&
+              state[:join_state] == 'steady'
+          end
         end
 
-        # Yields listeners in groups based on their subscription groups
-        # @yieldparam [Karafka::Routing::SubscriptionGroup] subscription group of given listeners
+        # Removes states that are no longer being reported for stopped/pending listeners
+        def evict
+          @mutex.synchronize do
+            @changes.delete_if do |_, details|
+              monotonic_now - details[:state_age_sync] >= @scale_delay
+            end
+          end
+        end
+
+        # Yields listeners in groups based on their subscription groups with multiplexing settings
+        # @yieldparam [Karafka::Pro::Routing::Features::Multiplexing::Config]
         # @yieldparam [Array<Listener>] listeners of a single subscription group
         def in_sg_families
           grouped = @listeners.group_by { |listener| listener.subscription_group.name }
@@ -307,7 +296,7 @@ module Karafka
             listener = listeners.first
 
             yield(
-              listener.subscription_group.name,
+              listener.subscription_group.multiplexing,
               listeners
             )
           end